[tb/update-locations] Rename and update locations for categorizer util
