+++ packages/perseus-core/src/index.ts
@@ -38,3 +38,4 @@ export type {ExpressionDefaultWidgetOptions} from "./widgets/expression";
 export type * from "./widgets/logic-export.types";
 
 export {default as getOrdererPublicWidgetOptions} from "./widgets/orderer/orderer-util";
+export {default as getCategorizerPublicWidgetOptions} from "./widgets/categorizer/categorizer-util";
