+++ packages/perseus-core/src/widgets/categorizer/categorizer-util.test.ts
@@ -1,4 +1,4 @@
-import getCategorizerPublicWidgetOptions from "./categorizer.util";
+import getCategorizerPublicWidgetOptions from "./categorizer-util";
 
 import type {PerseusCategorizerWidgetOptions} from "@khanacademy/perseus-core";
