+++ packages/perseus-core/src/widgets/categorizer/categorizer-util.ts
